Struct tzif::data::tzif::DataBlock

pub struct DataBlock {
    pub transition_times: Vec<Seconds>,
    pub transition_types: Vec<usize>,
    pub local_time_type_records: Vec<LocalTimeTypeRecord>,
    pub time_zone_designations: Vec<String>,
    pub leap_second_records: Vec<LeapSecondRecord>,
    pub standard_wall_indicators: Vec<StandardWallIndicator>,
    pub ut_local_indicators: Vec<UtLocalIndicator>,
}

A TZif data block consists of seven variable-length elements, each of which is a series of items. The number of items in each series is determined by the corresponding count field in the header. The total length of each element is calculated by multiplying the number of items by the size of each item. Therefore, implementations that do not wish to parse or use the version 1 data block can calculate its total length and skip directly to the header of the version-2+ data block.

In the version 1 data block, time values are 32 bits (TIME_SIZE = 4 bytes). In the version-2+ data block, present only in version 2 and 3 files, time values are 64 bits (TIME_SIZE = 8 bytes).

For consistency, this struct stores all time values in 64-bit integers even for version 1 data blocks.

The data block is structured as follows (the lengths of multi-byte fields are shown in parentheses):

     +---------------------------------------------------------+
     |  transition times          (timecnt x TIME_SIZE)        |
     +---------------------------------------------------------+
     |  transition types          (timecnt)                    |
     +---------------------------------------------------------+
     |  local time type records   (typecnt x 6)                |
     +---------------------------------------------------------+
     |  time zone designations    (charcnt)                    |
     +---------------------------------------------------------+
     |  leap-second records       (leapcnt x (TIME_SIZE + 4))  |
     +---------------------------------------------------------+
     |  standard/wall indicators  (isstdcnt)                   |
     +---------------------------------------------------------+
     |  UT/local indicators       (isutcnt)                    |
     +---------------------------------------------------------+

Fields
transition_times: Vec<Seconds>

A series of four- or eight-byte UNIX leap-time values sorted in strictly ascending order. Each value is used as a transition time at which the rules for computing local time may change. The number of time values is specified by the “timecnt” field in the header. Each time value SHOULD be at least -2**59.

(-2**59 is the greatest negated power of 2 that predates the Big Bang, and avoiding earlier timestamps works around known TZif reader bugs relating to outlandishly negative timestamps.)

transition_types: Vec<usize>

A series of one-byte unsigned integers specifying the type of local time of the corresponding transition time. These values serve as zero-based indices into the array of local time type records. The number of type indices is specified by the “timecnt” field in the header. Each type index MUST be in the range [0, “typecnt” - 1].

local_time_type_records: Vec<LocalTimeTypeRecord>

A series of LocalTimeTypeRecord objects.

time_zone_designations: Vec<String>

The string representations for a time-zone desigation, such as “PST” or “PDT”.

leap_second_records: Vec<LeapSecondRecord>

A series of LeapSecondRecord objects.

standard_wall_indicators: Vec<StandardWallIndicator>

A series of StandardWallIndicator objects.

ut_local_indicators: Vec<UtLocalIndicator>

A series of UtLocalIndicator objects.